Added new fonts page to the options dialog. Fixed some minor things with printing. Cleaned up vc7 warnings. Made all dates visible in the trial selection dialog on the calendar view.

+++ AgilityBook/src/Win/DlgRunScore.cpp
@@ -381,7 +381,7 @@ bool CDlgRunScore::GetText(CEdit* pEdit, short& val) const
 		return false;
 	CString str;
 	pEdit->GetWindowText(str);
-	val = atoi((LPCTSTR)str);
+	val = static_cast<short>(atoi((LPCTSTR)str));
 	return true;
 }
 bool CDlgRunScore::GetText(CEdit* pEdit, double& val) const
